+++ fungo/init.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/go-plugin"
 )
 
-const Version = "v0.5.3"
+const Version = "v0.5.4"
 
 var (
 	logger = Logger