+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## v0.5.4 (2024-01-18)
+
+- feat: add RunShell function to execute shell string commands and capture synchronous output
+- change: replace RunCommand with synchronous output
+
 ## v0.5.3 (2023-08-20)
 
 - feat: replace exec with myexec, add several exec helpers